Refactor CogAlgorithm to use CogAgentParameters struct for agent properties
This commit introduces a new CogAgentParameters struct to encapsulate
agent properties, replacing the previous use of #defines. It also
updates CogAlgorithm to use this new struct, improving code
maintainability and readability. The commit includes the addition of the
CogAgentParameters.h file and\
modifications to CogAlgorithm.h and CogAlgorithm.cpp to integrate the
new structure.